fix: quiz 위젯 클래스 및 토큰 미적용 수정
