+++ src/features/quiz/ui/blocks/QuizFooter.tsx
