+++ src/features/quiz/ui/story/QuizTeacher.stories.tsx
@@ -1,9 +1,9 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
-import QuizStudent from '../QuizStudent';
+import QuizTeacher from '../QuizTeacher';
 
 const meta = {
-  title: 'Feature/QuizStudent',
-  component: QuizStudent,
+  title: 'Feature/QuizTeacher',
+  component: QuizTeacher,
   argTypes: {
     question: {
       control: 'text',
@@ -12,6 +12,10 @@ const meta = {
     choices: {
       control: 'object',
       description: '선택지 목록',
+    },
+    correctIndex: {
+      control: 'number',
+      description: '정답 선택지의 인덱스 (0부터 시작)',
     },
     participantCount: {
       control: 'number',
@@ -25,29 +29,38 @@ const meta = {
       control: 'number',
       description: '정답률 (0~100, isEnded가 true일 때 표시)',
     },
-    isCorrect: {
-      control: 'boolean',
-      description: '학생의 정답 여부 (isEnded가 true일 때 표시)',
+    participantStatuses: {
+      control: 'object',
+      description: '선택지별 참여자 현황 목록',
     },
-    onSubmit: {
-      action: 'submitted',
-      description: '제출 버튼 클릭 시 호출되는 콜백 (선택한 인덱스 전달)',
+    onEnd: {
+      action: 'ended',
+      description: '퀴즈 종료 확인 시 호출되는 콜백',
     },
   },
-} satisfies Meta<typeof QuizStudent>;
+} satisfies Meta<typeof QuizTeacher>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
 const defaultChoices = ['파리', '런던', '베를린', '마드리드'];
 
+const defaultParticipantStatuses = [
+  { choiceIndex: 0, participants: ['김철수', '이영희', '박민준'] },
+  { choiceIndex: 1, participants: ['최지수'] },
+  { choiceIndex: 2, participants: ['정우진', '한소희'] },
+  { choiceIndex: 3, participants: ['강동원'] },
+];
+
 export const Active: Story = {
   args: {
     question: '프랑스의 수도는 어디인가요?',
     choices: defaultChoices,
-    participantCount: 12,
+    correctIndex: 0,
+    participantCount: 7,
     isEnded: false,
-    onSubmit: () => {},
+    participantStatuses: defaultParticipantStatuses,
+    onEnd: () => {},
   },
 };
 
@@ -55,25 +68,12 @@ export const Ended: Story = {
   args: {
     question: '프랑스의 수도는 어디인가요?',
     choices: defaultChoices,
-    participantCount: 24,
+    correctIndex: 0,
+    participantCount: 7,
     isEnded: true,
-    correctRate: 75,
-    correctAnswerIndex: 1,
-    isCorrect: true,
-    onSubmit: () => {},
-  },
-};
-
-export const EndedWrong: Story = {
-  args: {
-    question: '프랑스의 수도는 어디인가요?',
-    choices: defaultChoices,
-    participantCount: 24,
-    isEnded: true,
-    correctRate: 75,
-    correctAnswerIndex: 1,
-    isCorrect: false,
-    onSubmit: () => {},
+    correctRate: 43,
+    participantStatuses: defaultParticipantStatuses,
+    onEnd: () => {},
   },
 };
 
@@ -81,47 +81,37 @@ export const AllVariants: Story = {
   args: {
     question: '프랑스의 수도는 어디인가요?',
     choices: defaultChoices,
-    participantCount: 12,
+    correctIndex: 0,
+    participantCount: 7,
     isEnded: false,
-    onSubmit: () => {},
+    participantStatuses: defaultParticipantStatuses,
+    onEnd: () => {},
   },
   render: () => (
     <div className="flex flex-col gap-10">
       <div>
         <p className="body-medium text-black-500 mb-4">진행 중</p>
-        <QuizStudent
+        <QuizTeacher
           question="프랑스의 수도는 어디인가요?"
           choices={defaultChoices}
-          correctAnswerIndex={1}
-          participantCount={12}
+          correctIndex={0}
+          participantCount={7}
           isEnded={false}
-          onSubmit={() => {}}
+          participantStatuses={defaultParticipantStatuses}
+          onEnd={() => {}}
         />
       </div>
       <div>
-        <p className="body-medium text-black-500 mb-4">종료 - 정답</p>
-        <QuizStudent
+        <p className="body-medium text-black-500 mb-4">종료 후 결과 표시</p>
+        <QuizTeacher
           question="프랑스의 수도는 어디인가요?"
           choices={defaultChoices}
-          participantCount={24}
-          correctAnswerIndex={1}
+          correctIndex={0}
+          participantCount={7}
           isEnded
-          correctRate={75}
-          isCorrect
-          onSubmit={() => {}}
-        />
-      </div>
-      <div>
-        <p className="body-medium text-black-500 mb-4">종료 - 오답</p>
-        <QuizStudent
-          question="프랑스의 수도는 어디인가요?"
-          choices={defaultChoices}
-          participantCount={24}
-          correctAnswerIndex={1}
-          isEnded
-          correctRate={75}
-          isCorrect={false}
-          onSubmit={() => {}}
+          correctRate={43}
+          participantStatuses={defaultParticipantStatuses}
+          onEnd={() => {}}
         />
       </div>
     </div>
